Update the Libero core version for the 3 SmartHLS trainings for the 2022.3 release
The following cores were deprecated and need to be updated, otherwise the design
will fail to run with a fresh vault. See
https://confluence.microchip.com/pages/viewpage.action?pageId=415599015
- PF_CCC            -> 2.2.220
- PF_IOD_GENERIC_RX -> 2.1.110
- PF_DDR4           -> 2.5.109
- PF_SRAM_AHBL_AXI  -> 1.2.110

These cores are updated by changing the version specified in the script
`libero_flow.tcl` to use "*". From now on, these cores will always be
`download_core` and `create_and_configure_core` at the latest version.

All other cores are currently left unchanged at the current version. In the
future, if some of these cores were deprecated, we'll need to update our
scripts. The idea is that for this commit, we'll only change the smallest set of
cores that break the flow. If we update all cores to use the latest version,
there might be other unforeseen consequences that might require a lot of effort
to fix.

Testing: I tested locally on my Linux machine with Libero capture 2022_3_0_5_agni
and a fresh vault. The 3 trainings can run `libero_flow.tcl` correctly.