Fix testbench of wide_mult in Training3
In the last 'always' block of WideMultTopPf_Tb.sv, result checking is
performed when 'result' changes, and 'check_result_widemult' (the golden)
is supposed to change at the same time. In some simulation mode of ModelSim,
we encountered a situation that value change of both 'result' and
'check_result_widemult' happen at the same time, but 'check_result_widemult'
still holds the old value when simulation execution control goes into the
'always @(result)' block. To avoid this simulation implementation issue
affects the correctness of this testbench, an extra statement,
'@(negedge SYSCLK)' is added in the beginning of the 'always @(result)'
block body, to delay result checking to the next negedge clock to make sure
'check_result_widemult' is settled.