Fixed synthesis error for 2024_1 release
On Training3 some calculations were performed during the formal
port assignment or a module, potentially generating a wider bus than
the actual port definition.

Now the calculation is done in a separate wire assignment that performs
an automatic cast to the width of the wire to prevent the mismatch.

Probably Synplify got more strict about this kind of implicit type
casting during formal port assignment compared to the previous version.